installable_apps/requirements.txt
letta
pgserver~=0.1.4
py2app~=0.28.8
pystray
pillow
darkdetect
cutelog
#py2exe - not currently supporting 3.12 TODO upgrade or downgrade